meta: Update from upstream.

+++ src/web/vaev-base/z-index.h
@@ -3,9 +3,10 @@
 #include <karm-io/emit.h>
 
 #include "keywords.h"
+#include "numbers.h"
 
 namespace Vaev {
 
-using ZIndex = FlatUnion<Keywords::Auto, isize>;
+using ZIndex = FlatUnion<Keywords::Auto, Integer>;
 
 } // namespace Vaev
